+++ src/App.test.js
@@ -5,5 +5,6 @@ import App from './App';
 it('renders without crashing', () => {
   const app = shallow(<App />);
 
-  expect(app.find('.App').exists()).toEqual(true);
+  // expect(app.find('.App').exists()).toEqual(true);
+  expect(app).toMatchSnapshot();
 });
