+++ src/components/Login/Login.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Login from './Login';
+
+it('Login page renders without crashing', () => {
+  const login = shallow(<Login />);
+
+  expect(login).toMatchSnapshot();
+});
