+++ src/components/MyBudget/index.js
@@ -1,0 +1,3 @@
+import MyBudget from './MyBudget';
+
+export default MyBudget;
